off-by-one mismatch fix
